app/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
pydantic[email]
slowapi
limits
